Refactor tests to separate parallelism
